+++ src/imio/smartweb/core/contents/rest/events/endpoint.py
@@ -16,10 +16,23 @@ logger = logging.getLogger("imio.smartweb.core")
 
 
 class BaseEventsEndpoint(BaseEndpoint):
-    def __init__(self, context, request, fullobjects=0, batch_size=0):
+    def __init__(
+        self,
+        context,
+        request,
+        fullobjects=0,
+        batch_size=0,
+        sort_on=None,
+        sort_order=None,
+    ):
         self.fullobjects = fullobjects
         super(BaseEventsEndpoint, self).__init__(
-            context, request, fullobjects=fullobjects, batch_size=batch_size
+            context,
+            request,
+            fullobjects=fullobjects,
+            batch_size=batch_size,
+            sort_on=sort_on,
+            sort_order=sort_order,
         )
 
     def __call__(self):
@@ -95,7 +108,6 @@ class BaseEventsEndpoint(BaseEndpoint):
             "metadata_fields=has_leadimage",
             "metadata_fields=UID",
             "metadata_fields=language",
-            "sort_on=event_dates",
             "fullobjects={}".format(self.fullobjects),
         ]
         if self.batch_size == 0:
@@ -105,6 +117,7 @@ class BaseEventsEndpoint(BaseEndpoint):
         if self.context.selected_event_types is not None:
             for event_type in self.context.selected_event_types:
                 params.append(f"event_type={event_type}")
+        params += self.sort_params("event_dates")
         params = self.construct_query_string(params)
         url = f"{EVENTS_URL}/{self.remote_endpoint}?{params}"
         if is_log_active():
@@ -128,9 +141,22 @@ class EventsEndpointGet(BaseService):
     def reply(self):
         return EventsEndpoint(self.context, self.request)()
 
-    def reply_for_given_object(self, obj, request, fullobjects=0, batch_size=0):
+    def reply_for_given_object(
+        self,
+        obj,
+        request,
+        fullobjects=0,
+        batch_size=0,
+        sort_on=None,
+        sort_order=None,
+    ):
         return EventsEndpoint(
-            obj, request, fullobjects=fullobjects, batch_size=batch_size
+            obj,
+            request,
+            fullobjects=fullobjects,
+            batch_size=batch_size,
+            sort_on=sort_on,
+            sort_order=sort_order,
         )()